postsManagerPage에 있던 addPost, updatePost 함수 분리
